:recycle: Refactor command execution and error handling in CLI wrapper
