+++ wrapper/gh/list_tags.ts
@@ -1,4 +1,4 @@
-import { exec } from "./exec.ts";
+import { execWithRetry } from "../exec.ts";
 import type { GitHubCliOptions } from "./types.ts";
 
 /**
@@ -55,5 +55,5 @@ export async function listTags(
   if (host) args.push("--hostname", host);
   if (jq) args.push("-q", jq);
 
-  return await exec(args);
+  return await execWithRetry("gh", args);
 }
