+++ wrapper/gh/mod.ts
@@ -1,4 +1,3 @@
-export { GitHubCliError } from "./exec.ts";
 export * from "./graphql.ts";
 export * from "./list_commits.ts";
 export * from "./list_tags.ts";
