+++ wrapper/git/mod.ts
@@ -1,4 +1,3 @@
-export { GitError } from "./exec.ts";
 export * from "./describe.ts";
 export * from "./list_remotes.ts";
 export * from "./rev_parse.ts";
