+++ wrapper/mod.ts
@@ -1,0 +1,1 @@
+export { CliError } from "./exec.ts";
